# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee CRUD API",
        "version": "1.0.0",
        "description": "A simple Employee CRUD API with AWS Secrets Manager integration"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
        "/employees": {
            "get": {
                "summary": "Get all employees",
                "tags": [
                    "Employees"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new employee",
                "tags": [
                    "Employees"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "salary": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "name",
                                    "position"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "summary": "Get employee by ID",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "summary": "Update an employee",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "salary": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an employee",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Employee deleted"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        },
        "/debug/env-secrets": {
            "get": {
                "summary": "Get secrets loaded in process.env",
                "tags": [
                    "Debug"
                ],
                "responses": {
                    "200": {
                        "description": "JSON object with secrets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "DB_HOST": {
                                            "type": "string"
                                        },
                                        "API_KEY": {
                                            "type": "string"
                                        },
                                        "PORT": {
                                            "type": "string"
                                        },
                                        "AWS_SECRET_NAME": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/debug/refresh-secrets": {
            "post": {
                "summary": "Manually refresh secrets from AWS",
                "tags": [
                    "Debug"
                ],
                "responses": {
                    "200": {
                        "description": "Secrets refreshed successfully"
                    },
                    "500": {
                        "description": "Failed to refresh secrets"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Employee": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "number"
                    }
                }
            }
        }
    }
}